Implement results saving
Result of each inference is write to csv file which is
located in "/data/data/<package-name>/files" directory.

In order to view result file, in Android Studio go to View->
Tool Windows->Device File Explorer and then navigate to
directory which was mentioned earlier. If file is not present
try to right click on directory and then click Synchronize.